Refactor to move research unlock logic to a separate module
